requirements.txt
Brotli==1.1.0
colorama==0.4.6
glfw==2.9.0
imgui==2.0.0
inflate64==1.0.3
multivolumefile==0.2.3
numpy==2.3.0
psutil==7.0.0
py7zr==1.0.0
pybcj==1.0.6
pycryptodomex==3.23.0
PyOpenGL==3.1.9
pyppmd==1.2.0
pyzstd==0.17.0
texttable==1.7.0
typing_extensions==4.14.0
